Avoid integer vs unsigned integer error.

+++ src/features/mm_bone_data_feature.cpp
@@ -46,7 +46,7 @@ uint32_t MMBoneDataFeature::get_dimension_count() const {
 
 PackedFloat32Array MMBoneDataFeature::bake_animation_pose(Ref<Animation> p_animation, float time) const {
     PackedFloat32Array result;
-    for (size_t i = 0; i < bone_names.size(); ++i) {
+    for (int64_t i = 0; i < bone_names.size(); ++i) {
         const String bone_path = String(_skeleton_path) + String(":") + bone_names[i];
 
         const int32_t pos_track = p_animation->find_track(bone_path, Animation::TrackType::TYPE_POSITION_3D);
@@ -81,7 +81,7 @@ PackedFloat32Array MMBoneDataFeature::bake_animation_pose(Ref<Animation> p_anima
 
 PackedFloat32Array MMBoneDataFeature::evaluate_runtime_data(const MMQueryInput& p_query_input) const {
     PackedFloat32Array result;
-    for (size_t i = 0; i < bone_names.size(); ++i) {
+    for (int64_t i = 0; i < bone_names.size(); ++i) {
 
         const Vector3& pos = p_query_input.skeleton_state.find_bone_state(bone_names[i]).pos;
         const Quaternion& rot = p_query_input.skeleton_state.find_bone_state(bone_names[i]).rot;
